requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
openai
python-docx
python-pptx
python-dotenv
email-validator
